Rename: 'models' 모델 파일명 변경 -> 조금 더 직관적으로
